requirements.txt
Flask==0.9
Flask-Login==0.1.3
Flask-OpenID==1.1.1
Flask-SQLAlchemy==0.16
Flask-Sijax==0.3.1
Flask-WTF==0.8.3
Jinja2==2.6
SQLAlchemy==0.7.9
Sijax==0.2.6
Tempita==0.5.1
WTForms==1.0.4
Werkzeug==0.8.3
argparse==1.2.1
decorator==3.4.0
logilab-astng==0.24.1
logilab-common==0.59.0
pylint==0.26.0
python-openid==2.2.5
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
